requirements.txt
asgiref==3.2.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.1
djangorestframework==3.11.0
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.2
numpy==1.17.4
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7